requirements.txt
alembic==1.15.1
annotated-types==0.7.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
soupsieve==2.5
greenlet==3.1.1
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
SQLAlchemy==2.0.39
typing_extensions==4.12.2
python-dotenv>=0.21.0
passlib==1.7.4
fastapi==0.115.1
uvicorn==0.34.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic_ai==0.1.2
stripe==9.7.0
aiosmtplib==3.0.2 
jinja2==3.1.3
celery==5.3.6
redis==5.0.4
pillow==11.3.0
sendgrid== 6.12.4 